awadawd

+++ setup.py
@@ -1,23 +1,37 @@
 #TODO Автоматическая установка Maten
 from pyrogram import Client
 from pyrogram.errors import SessionPasswordNeeded
-from main import restart
 import os
+import sys
+import asyncio
 from getpass import getpass
 
-async def setup(client: Client):
-    await client.connect()
+def restart():
+    python = sys.executable
+    os.execl(python, python, "main.py")
+
+async def setup():
     print("Hello! Welcome to Maten userbot setup. Please select your language: [EN/ru]")
     lang = input("Language: ").strip().lower() == 'ru'
+
+    api_id = input("Введите API ID: " if lang else "Enter API ID: ")
+    api_hash = input("Введите API HASH: " if lang else "Enter API HASH: ")
+    
+    with open("config.ini", "w") as f:
+        f.write(f"[pyrogram]\napi_id = {api_id}\napi_hash = {api_hash}\n")
+
+    client = Client("maten", api_id=api_id, api_hash=api_hash)
+    await client.connect()
+
     if lang:
         phone_number = input("Введи номер телефона: ")
         if not phone_number:
-            os.remove("maten.session")
+            if os.path.exists("maten.session"): os.remove("maten.session")
             return
         sent_code = await client.send_code(phone_number)
         code = input("Введи код, который пришел тебе в Telegram: ")
         if not code:
-            os.remove("maten.session")
+            if os.remove("maten.session"): os.remove("maten.session")
             return
         try:
             user = await client.sign_in(
@@ -36,12 +50,12 @@ async def setup(client: Client):
         print("Please enter your phone number: ")
         phone_number = input("Phone number: ")
         if not phone_number:
-            os.remove("maten.session")
+            if os.path.exists("maten.session"): os.remove("maten.session")
             return
         sent_code = await client.send_code(phone_number)
         code = input("Enter the code you received in Telegram: ")
         if not code:
-            os.remove("maten.session")
+            if os.path.exists("maten.session"): os.remove("maten.session")
             return
         try:
             user = await client.sign_in(
@@ -57,4 +71,6 @@ async def setup(client: Client):
         restart()
 
 if __name__ == "__main__":
-    os.remove("maten.session")
+    if os.path.exists("maten.session"):
+        os.remove("maten.session")
+    asyncio.run(setup())
